追加1

+++ album/views.py
@@ -2,7 +2,6 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import Image
 from .forms import ImageForm
 
-
 def upload(request):
     if request.method == "POST":
         form = ImageForm(request.POST, request.FILES)
@@ -27,21 +26,29 @@ def betails(request):
     return render(request, 'album/betails.html', context)
 
 def edit(request, id):
+    image = get_object_or_404(Image, pk=id)
     if request.method == "POST":
-        image = get_object_or_404(Image, pk=id)
-        form = ImageForm(request.POST, request.FILES)
+        form = ImageForm(request.POST, instance = image)
         if form.is_valid():
-            form.save()
+            image=form.save(commit=False)
+            image.picture=request.FILES['picture']
+            image.save()
             return redirect('album:showall')
+    imageform = ImageForm(instance = image)
     context = {
         'image':image,
-        'imageform': imageForm
+        'imageform': imageform
     }
     return render(request, 'album/edit.html', context)
-    
+
 def delete(request, id):
     image = get_object_or_404(Image, pk=id)
-    image.delete()
-    imageForm = ImageForm(instance=image)
-    context = {'image':image}
-    return render(request, 'album/betails.html', context)
+    if request.method == "POST":
+            image.delete()
+            return redirect('album:showall')
+    imageform = ImageForm(instance = image)
+    context = {
+        'image':image,
+        'imageform': imageform
+    }
+    return render(request, 'album/delete.html', context)
